+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 [WHATWG URL specification]: https://url.spec.whatwg.org/#file-slash-state
 
+* When a root-relative URLs like `/foo` is converted to a Windows path using
+  `fromUrl()`, it is now resolved relative to the drive letter. This matches
+  IE's behavior.
+
 ## 1.4.0
 
 * Add `equals()`, `hash()` and `canonicalize()` top-level functions and